// src/types/ProjectFormData.d.ts
export interface ProjectFormData {
  docId: string;
  id: number;
  projectName: string;
  description: string;
  thumbnail: string;
  client: string;
  startDate: string;
  endDate: string;
  teamSize: string;
  planning: string;
  design: string;
  publishing: string;
  development: string;
  badgeProjectDevice: string;
  badgeProjectType: string;
  badgeParticipation: string;
  demoUrl: string;
  githubUrl: string;
  canvaUrl: string;
  figmaUrl: string;
  swaggerUrl: string;
  techStack: string[];
  descriptionDetail: string;
}